+++ ostap/fitting/tests/test_fitting_models.py
@@ -1080,6 +1080,57 @@ def test_rasingcosine () :
 
 
 # =============================================================================
+## Breit-Wigner with interference
+# =============================================================================
+## def test_bwi () :
+if 1 < 2 :     
+    logger.info ('Test BWI_pdf' )
+    
+    ff = cpp.Ostap.Math.FormFactors.BlattWeisskopf( 1 , 3.5 ) ## formfactor 
+    bw = cpp.Ostap.Math.BreitWigner (
+    m.value() ,
+    m.error() ,
+    0.150     , ## m1 
+    0.150     , ## m2 
+    1         , ## orbital momentum
+    ff          ## formfactor 
+    )
+    
+    model = Models.Fit1D (
+        signal = Models.BWI_pdf
+        ( name        = 'BWI'             ,
+          breitwigner = bw                ,     
+          xvar        = mass              ,
+          mean        = signal_gauss.mean ,
+          bkg         = -1                ) , 
+        background = Models.PolyPos_pdf ('BkgBWI', xvar = mass , power = 1 )) 
+    
+    signal = model.signal 
+    model.S.setVal(5000)
+    model.B.setVal(500)
+    
+    signal.mean.fix ( m.value() )
+    
+    with rooSilent() : 
+        result, frame = model. fitTo ( dataset0 )
+        signal.mean .release()
+        signal.gamma.release()
+        result, frame = model. fitTo ( dataset0 )
+
+    if 0 != result.status() or 3 != result.covQual() :
+        logger.warning('Fit is not perfect MIGRAD=%d QUAL=%d ' % ( result.status() , result.covQual()  ) )
+        print(result)
+    else :
+        logger.info ( 'Signal & Background are: %-28s & %-28s ' % ( result ( 'S'         )[0] , result( 'B'           )[0] ) ) 
+        logger.info ( 'Mean                 is: %-28s ' %  result ( signal.mean  )[0] )
+        logger.info ( 'Gamma                is: %-28s ' %  result ( signal.gamma )[0] )
+        logger.info ( "Breit-Wigner function\n%s" % result.table ( prefix = "# " ) )
+
+    models.add ( model )
+
+
+
+# =============================================================================
 ## check that everything is serializable
 # =============================================================================
 def test_db() :
@@ -1095,38 +1146,40 @@ def test_db() :
 # =============================================================================
 if '__main__' == __name__ :
 
-    test_gauss          () ## simple Gaussian PDF                       + background 
-    test_crystalball    () ## Crystal Ball                              + background
-    test_crystalball_RS () ## right-side Crystal Ball                   + background  
-    test_crystalball_DS () ## double side Crystal Ball                  + background 
-    test_needham        () ## Needham function (CB with alpha=f(sigma)) + background 
-    test_apollonios     () ## Apollonios function                       + background 
-    test_apollonios2    () ## modified Apollonios function              + background 
-    test_bifurcated     () ## bifurcated Gaussian function              + background 
-    test_2gauss         () ## double     Gaussian function              + background 
-    test_gengauss_v1    () ## generalized Gaussian function V1          + background 
-    test_gengauss_v2    () ## generalized Gaussian function V2          + background 
-    test_skewgauss      () ## skew gaussian                             + background 
-    test_qgauss         () ## q-Gaussian function                       + background 
-    test_bukin          () ## Bukin - skew Gaussian core with exponential tails  + background 
-    test_studentT       () ## Student-t shape                           + background 
-    test_bifstudentT    () ## Bifurcated Student-t shape                + background 
-    test_sinhasinh      () ## Sinh-Asinh distribution                   + background 
-    test_johnsonSU      () ## Johnson-SU distribution                   + background 
-    test_atlas          () ## Modified Gaussian used by ATLAS/Zeus      + background 
-    test_sech           () ## Sech (1/cosh)  distribution               + background 
-    test_losev          () ## Asymmetric hyperbilic secant distribution + background 
-    test_logistic       () ## Logistic distribution                     + background 
-    test_voigt          () ## Voigt profile                             + background 
-    test_pvoigt         () ## Pseudo-Voigt(approximation to Voigt)      + background 
-    test_bw             () ## Breit-Wigner(+resolution)                 + background 
-    test_slash          () ## Slash-function                            + background 
-    test_rasingcosine   () ## Raising Cosine                            + background 
-    test_laplace        () ## Laplace-function                            + background 
+    ## test_gauss          () ## simple Gaussian PDF                       + background 
+    ## test_crystalball    () ## Crystal Ball                              + background
+    ## test_crystalball_RS () ## right-side Crystal Ball                   + background  
+    ## test_crystalball_DS () ## double side Crystal Ball                  + background 
+    ## test_needham        () ## Needham function (CB with alpha=f(sigma)) + background 
+    ## test_apollonios     () ## Apollonios function                       + background 
+    ## test_apollonios2    () ## modified Apollonios function              + background 
+    ## test_bifurcated     () ## bifurcated Gaussian function              + background 
+    ## test_2gauss         () ## double     Gaussian function              + background 
+    ## test_gengauss_v1    () ## generalized Gaussian function V1          + background 
+    ## test_gengauss_v2    () ## generalized Gaussian function V2          + background 
+    ## test_skewgauss      () ## skew gaussian                             + background 
+    ## test_qgauss         () ## q-Gaussian function                       + background 
+    ## test_bukin          () ## Bukin - skew Gaussian core with exponential tails  + background 
+    ## test_studentT       () ## Student-t shape                           + background 
+    ## test_bifstudentT    () ## Bifurcated Student-t shape                + background 
+    ## test_sinhasinh      () ## Sinh-Asinh distribution                   + background 
+    ## test_johnsonSU      () ## Johnson-SU distribution                   + background 
+    ## test_atlas          () ## Modified Gaussian used by ATLAS/Zeus      + background 
+    ## test_sech           () ## Sech (1/cosh)  distribution               + background 
+    ## test_losev          () ## Asymmetric hyperbilic secant distribution + background 
+    ## test_logistic       () ## Logistic distribution                     + background 
+    ## test_voigt          () ## Voigt profile                             + background 
+    ## test_pvoigt         () ## Pseudo-Voigt(approximation to Voigt)      + background 
+    ## test_bw             () ## Breit-Wigner(+resolution)                 + background 
+    ## test_slash          () ## Slash-function                            + background 
+    ## test_rasingcosine   () ## Raising Cosine                            + background 
+    ## test_laplace        () ## Laplace-function                            + background 
     
-    ## check finally that everything is serializeable:
-    test_db ()          
+    ## ## check finally that everything is serializeable:
+    ## test_db ()
+    pass 
+         
 
 # =============================================================================
-# The END 
+##                                                                      The END 
 # =============================================================================