+++ ReleaseNotes/release_notes.md
@@ -26,7 +26,7 @@ canvas >> 'plots.txz' ## make plot in all formats and store them in tar/LZMA  fo
   1. add SciPy-based bspline interpolation for functions `ostap/math/sp_interpolation.py'
   1. imporve pseudo-abstract operations `ostap/math/operations.py'
   1. extend interpolation tests
- 
+  1. add simple shapes to probe signal/background interference `Ostap::Models::BWI` and `BWI_pdf` 
  
 ## Backward incompatible changes